requirements.txt
pandas
numpy
requests
scikit-learn
xgboost
ta
streamlit
python-dotenv
joblib
matplotlib
plotly